Arousal-fix-fixed
Made on_species_gain do the arousal check and apply it properly. Moved
equip to happen AFTER the ckey is assigned, as I couldn't check the
clients prefs before that and it just defaulted to TRUE.
Problem: Human ghost-roles are already human and would not use
on_species_gain. I had 2 fixes for this: Either I made them do so
anyway, doing set_species with human. This is probably redundant and
might break things in some unexpected way....
2nd possible fix is what I am pushing now, just making an else statement
for that situation.