Captain's Laser accessible by captain
Changes the access required from centcomm special ops to captain access.
Comment says it was on purpose, but seems poorly explained. As it is, a
lot of people don't seem to know display cases even CAN be opened. And
the gun is the captain's, why would it need special ops access to
retrieve?