Solves Ninja-tech theft
You can now steal tech from the R&D console directly.
You can STILL steal it using a disk, by hitting your suit with the disk,
but people don't seem to use that. This code is horrible, by the way, it
does nothing to fix the fact that it's a copy-paste for both the server
and the console.