Cloning-Genitalia-Fix (#277)
* Cloning-Genitalia-Fix

Fixes the issue with genitals not being re-created with cloning. This
should make it so other methods of coming back also gives the organs
your DNA says you should have, such as podding people.

* Sanity checks and fixes

Removed the OLD sanity check for drawing eyes. It was refusing to draw
eyes if you had ANY disability. 's far as I know, blind people still
have eyes and so do mute people.
Also, put a sanity check in give_genitals, just in case a species has
DNA with genitals but is a NOGENITALS species. Somehow. IT COULD HAPPEN,
ALRIGHT?! ^^

* TRAVIS LOVE ME AGAIN

This commit fixes the problem that made Travis hate me previously. Even
though THAT part of the code had not been changed...
Anyway, it looks more neat now, so that's good.