Huge Sprite Rework (#342)
* Revert "Revert "Huge Sprite Rework""

* Compilation of all feedback fixes

This commit contains all the suggested fixes so far and the changes we
came up with through discussion about the code.
For example, genitals **may** now use alternate icons for aroused
states, but does not have to, this is decided by a boolean variable in
the sprite accessory itself, defaulting to not using alternate icons.
Genital icon_state names now follow a new format to match these new
options in a modular way for all kinds of genitals.

* Conflict solving

reformats all the taurs to use the TAUR layer instead of the BODY_FRONT
layer. Same for the new Eevee sprites.

* Committing the old, unfixed file.

This will cause errors if used, but at least it won't cause conflicts.

* Contains fixed sprite names

The shepherd taur body should STILL have a hole in the body due to a
pixel error.
I have forgotten which pixes these are, so.... **le shrug**

* Anger commit

Commiting this in anger. I WILL make it work damnit!

* Proper name fixes.

Also renames the taur cow to no longer have an incorrectly duplicate
icon_state.

* Shepherd taur pixel fix

Fixes the missing pixels from the taur body that caused a hole to appear
when viewed from the sides.

* Conflict resolution

This PR will cause some sprite-breaking overlay errors.
But it's the only way to handle the conflicting .dmi.
So I'll make a separate PR to fix the sprites later.

* Fixes global lists

Now they use whatever weird new system we got for them.