Implementing feedback fixes
Also moved the procs for genital size-change to the genitals themselves.
Breasts code refactored to behave more like other genitals do with
sizes, i.e using numbers for size and not strings.
Defines are also used for this, so the max breast size define has been
updated to reflect the current size restrictions.