Adding citadel chems and flufftext
Aphrodisiac and Anaphrodisiac added to borer chem list. Now that TG
removed borers, maybe they'll stay on this time.
Also added flufftext for borer_chems, so that there is some kind of
feedback to the host when their borer puts something in them. I'm not
sure what span class should be used, but the chem is usually beneficial
and the only chem that might be dangerous (meth) is not very subtle as
to miss it just because the span class...

Also added snowflake exception for citadel chem flavor text, because
some people/species cannot feel arousal and thus would not notice the
chemical having any effect.