Finished testing
Added the Late-assassinate target objective, made it have a random timer
when it checks for a target and a random % chance to just skip it for a
cycle.
ALSO fixed the normal find_target_by_role proc, as it used to not be
random at all, meaning everyone who found a target using that proc would
target the same poor person. Now it's randomly selected from all
available targets.
This is an indirect nerf to lings I suppose, as TEAM lings can now roll
different targets, but if we WANT that, then it should've been coded on
the ling-level, not on the normal objective code.