Added flared taur penises
FIRST STEP ONLY DO NOT PULL REQUEST THIS!!!!
The files are still messy and the penis is just another shape to choose
from right now for testing purposes!
Flipped the order of GENITALS_BEHIND_LAYER and BODY_BEHIND_LAYER, as
taurs need to be drawn on top of the genitalia, and I don't think
anything else uses this that would need genitals drawn on top of them.