Fixes on-ground sprites
The OLD bug barbed knotted cocks being invisible when not inside a body
is fixed.
This fix solves why taur cocks were invisible when cut off or gibbed the
body.
This also removes the lowershape() workaround that was being used.
I'll get around a bigger refactor of other genitalia in another PR, as
the scope of this was just to add taur cocks to begin with...