Flavortext transition finished
Now properly moves OLD flavortext to the new format and then removes the
old format.
If no old format is found, it defaults to loading the new format.
Flavortext is now stored in your DNA, and is loaded into your characters
flavor_text variable upon creating the character or messing with the
DNA.
This allows cloners/pods/changelings to use the correct flavortext.