Sets all food prefs to NONE
People can RP their food likes/dislikes. This stuff, like deciding what
is/isn't edible to slimes and such is unasked for and unecessary.
Also, they didn't address that this stealth-nerfs non-humans as
donk-pockets and such food items were put in a lot of places, expecting
to be used for healing, but can now not be eaten by, for example,
lizards.