Rebalanced pod people so they arnt as weak.
This will allow podpeople to regen at a steady pace without them being able to take tool boxes to the head like candy. it takes around 15 seconds to notably heal compared to the 40 it does just for a little bit and the 3 from before it was nerfed. The fact that podpeople need to go to the medbay to fix burn damage made me decide to make their brute regen stronger, not to mention that a single welder smach gives 18.75 damage and a shot to the leg with a laser carbine does 25 burn damage.