Fixed the drying rack for real this time
This is a very tricky issue as tgstation has the same code but I have no
clue how the hell they're not getting the same error if -1 is supposed
to return false all the time. There was some committ on our end that
caused this issue but I have no clue what it is but this at least fixes
it without runtime errors.