Sake! (#5538)
* Sake!

:cl: Fel and LeonDuvall
add: You can now drink sake! Can be found in the bar, or made with rice
and sugar.
/:cl:

Sprites and idea courtesy of LeonDuvall.
Can be made with 15u rice and 5u sugar at 400k.

* Does the thing

Modularization of vending machines would require a lot more effort than
is remotely worth for such a simple change. Fortunately, not a lot of
new alcohols are added to it often, so it should be fine.