The Open Galaxy
:cl: Ah yes "Reapers"
add: Finally adds turians, but as a body marking, to have a better look, select the mammal, remove the snout and you won't have any weird looks, you can still add ears and tail, even taur body if anyone is "unique" enough to use it.

:cl: Scolipede, i choose you!!
add: Adds a new whole specie! Scolipede have snout, body marking, taur, tail and ears, for a more good look, try to find dark red colors to fit with it.

add: Added a bit news things, such as bee body marking, insect tail, fish tail and gills which is found in ears, horn for anyone that want to make a demoness or goat, added a tentacle tail, which i can see it will be mostly be used for slimes, the tentacle tail itself was an old sprite of citadel, i only brought back since i found a place to fit it, the real creator of the sprite will have the credit he/she deserves.

fix: Fixed the sprite bug of the octopus taur, it also had it name changed to tentacle
/:cl: