Update revenant_names.json
Removes the chance of revenant to roll in the name so as to slowly shift their names from Rev, to mainly phantom. Added another 1 phantoms in the list to give it a higher roll chance. No more "Rev in medbay!" causing unneeded revolutionary preparations as well as trying to give them a more generalized name of phantom.