Ninja Rebalancing
Ninjas now spawn with Sleeping Carp. To compensate for this, the cooldown of their sword's dash ability has been increased, and the energy net has been changed to only affect people who are unconscious, IE critted or sleeping; or stamcritted, to account for being incapacitated by a baton or other method.

This should adjust the current meta of security using shotguns against nearly any threat against them, while also removing the current meta of ninjas grabbing a bag of holding and solely relying on weapons instead of their abilities and melee combat. Testmerge for now, until more data can be gathered on whether this is a good change or not.