Doubles stamina regen when standing/laying down
Does just as it says on the tin, doubles stamina regeneration when standing/laying down. Keeps hard stam crit regen as buffing that would make it harder/difficult to cuff. 

This PR is going to have to be playtested as it's not particularly easy to locally test. Testmerging and tweaking it w/ observed input would be preferred.