Ports Xenobio Console Hotkeys from TG
## About The Pull Request
This enables xenobiologists to perform all of the (unlocked) slime console functions by simply holding down a modifier key and primary-clicking a slime/monkey/empty floor space. This functionality has been ported from TG and modified to work for Cit.

## List of Changes
* Button added to slime console that prints out the hotkey commands to the user. [NEEDS BUTTON ICON]
* Shift-click a slime to pick it up, or the floor to drop all held slimes. (Requires Basic Slime Console upgrade)
* Ctrl-click a slime to scan it.
* Alt-click a slime to feed it a potion. (Requires Advanced Slime Console upgrade)
* Ctrl-click on a dead monkey to recycle it, or the floor to place a new monkey. (Requires Monkey Console upgrade)
* If the console does not have the required upgrade, an error message will print to the user.
* When recycling monkeys, the console will now report how many monkeys it has (will not report decimal increases).
* Console now alerts you when you're out of monkeys and reports your current decimal amount.
* Console messages are now styled consistently.
* You can now pick up a single slime from a pile, instead of all of them at once.

 ## Why It's Good For The Game
Adds QoL features for Xenobiologists and increases department efficiency. Removes the frustration of trying to line up a hyper-sensitive camera reticle on a moving slime with your keyboard, especially for users with laggy clients.

## Notes
This PR may contain bugs and poor coding practices.
TG allows for consoles to be linked to a physical monkey recycler, sharing its contents across all 3 machines. This PR does NOT port that functionality because I'm lazy and don't know what I'm doing.
Special thanks to everyone in #development-discussions-main for helping me with my dumb mistakes and to @Studious Cirno#4894 for suggesting this PR.