Swaps syringe with a sleep pen in poision kit
The poison kit is in a weird spot at the moment, because the syringe that it comes with means that anyone who gets injected with /any/ of the chemicals instantly screams that somoene is injecting them, and that they are a traitor.
Increases the cost of the poision kit from 6 to 8 TC as compensation.