[Balance Pass] - Supermatter Mega Delam (Tesla)
All I'm doing here is raising the power_penalty_thresholds to a higher number so that it takes more time and effort to force the mega delamination.

Also by leaving a wider gap between Severe and Critical stages, it gives more warning to AI and Engineers that Tesla delamination is happening soon.

As some of us know by now, it's /very/ easy to force Tesla delams and that uhh... it's aids to deal with due to all the effects happening so quickly as it powers through all the stages from just 3 simple switches.

It takes effort and time for Singulo delam to start happening due to the high amount of moles needed on. that. tile. 

Does not take much for Tesla. Fixing that now.