Made coolant gases stronger for supermatter.
This is to make up for unforeseen consequences of having the gases update more often: rather than gases moving from one side of the supermatter to the other, they spread out instantly, which makes the heat spread much faster. To counteract this, I'm making coolant gases significantly stronger. This may need to be dialed back, but for now, it's a fix.