Fuck it, Lambda Station, here we go!
## About The Pull Request
Welcome to Lambdastation! After almost two years in developement, the map is complete. Starting as a community project on Oracle station in which different players worked in game to create the map, headed by AndrewMontagne who created code to save the project as it was being worked on. After work on cargo, escape, medical, and parts of engineering, the project ended and I asked for the file.

I continued working on the file for a few months before jumping to a newer tg branch, in which I had to recreate the map from scratch. After another move to continue working the map on Citadel code, I have finally finished the map!

I'd like to thank Sishen for converting this to tgm formatting, and the original team of players that worked to shape the map before I finished it.

## Why It's Good For The Game
New map, with a large maintenance system that allows additional rooms to be made, along with genpop! Good for mid to high population.

## Changelog
:cl:
add: Lambdastation and it's accompanying files
/:cl: