rites actually work now
Rites can be casted now, added a show info option on the sigil of rites, adds a TON of debug messages, if this is PRd while they are still in yell at me loudly. Also adds the rite of advancement, a rite used to implant organs into a servant, without surgery being required.