Makes the Sigil of rites summonable and adds some more rites
The sigil of rites can now be summoned for 1400 power at tier two.
Two new rites are available: The rite of woundmending, which heals any wounds (not damage) a target may have, but causes 10 toxins damage per wound healed aswell as the rite of the claw, which summons a brass claw implant on the rune. This rite specifically can only be used a limited amount of times